requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==3.0.0
langchain==0.0.225
flask
pypdf
python-dotenv
-e .